+++ source/ApiInterface.h
@@ -1,3 +1,22 @@
+/*
+# Playstation 3 Cloud Drive
+# Copyright (C) 2013-2014   Mohammad Haseeb aka MHAQS
+# Copyright (C) 2023        Vitaly Hodiko aka vitaly.x
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program. If not, see <http://www.gnu.org/licenses/>.
+#
+*/
 #pragma once
 
 #include <string>
@@ -10,11 +29,15 @@
 #include "CurlAgent.h"
 #include "OAuth2.h"
 
+#include <vector>
+
 class APIInterface {
    public:         
        double m_prog_func_last_progress;
 
-       const std::string *m_api_key_map;
+       std::vector<std::string> m_api_key_map;
+         
+       const std::string m_api_root_url;
        const std::string m_client_id;
        const std::string m_client_secret;
        const std::string m_token_url;
@@ -25,20 +48,21 @@ class APIInterface {
        Json m_remote_resource_root;
 
         APIInterface(
+                std::string _api_url, 
                 std::string _client_id, 
                 std::string _client_secret,
                 std::string _token_url, 
                 std::string _device_url
         ):
-
+                m_api_root_url(_api_url),
                 m_client_id(_client_id),
                 m_client_secret(_client_secret),
                 m_token_url(_token_url),
                 m_device_url(_device_url)
         {}
 
-        void init(){
-                m_auth_token = new OAuth2(m_token_url, m_device_url, m_client_id, m_client_secret, m_api_key_map);
+        void init() {
+                m_auth_token = new OAuth2(m_token_url, m_device_url, m_client_id, m_client_secret, &m_api_key_map);
         }
 
         bool ShouldRetryCheck( long response );