+++ source/GoogleApiImpl.h
@@ -1,3 +1,22 @@
+/*
+# Playstation 3 Cloud Drive
+# Copyright (C) 2013-2014   Mohammad Haseeb aka MHAQS
+# Copyright (C) 2023        Vitaly Hodiko aka vitaly.x
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program. If not, see <http://www.gnu.org/licenses/>.
+#
+*/
 
 #include <string>
 #include "Json.h"
@@ -11,17 +30,25 @@ class GoogleApiImpl: public APIInterface {
    public:
         GoogleApiImpl() : 
             APIInterface(
+                "https://www.googleapis.com/",
                 GOOGLE_API_ID, 
                 GOOGLE_API_KEY,
                 "https://www.googleapis.com/oauth2/v4/token",
                 "https://accounts.google.com/o/oauth2/device/code"
             )
         {
-            const std::string _map[E_NUM] = {
-                    /* DEVICE_CODE_KEY */ "code",
-                    /* GRANT_TYPE_VALUE */ "http://oauth.net/grant_type/device/1.0",
+            /* scope
+                http.Escape("email profile ") +*/ 
+            // http.Escape("https://www.googleapis.com/auth/userinfo.email") + http.Escape(" ") +
+            // http.Escape("https://docs.google.com/feeds") + http.Escape(" ") + // Deprecated, but still working
+            // http.Escape("https://www.googleapis.com/auth/userinfo.profile");*/
+            m_api_key_map = {
+                    /* DEVICE_CODE_KEY */           "code",
+                    /* GRANT_TYPE_VALUE */          "http://oauth.net/grant_type/device/1.0",
+                    /* DEVICE_GRANT_SCOPE_VALUE */  "https://www.googleapis.com/auth/drive.file",
+                    /* DEVICE_CODE_USER_URL */      "verification_url"
             };
-            m_api_key_map = _map;
+            
             init();
         }